backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
faiss-cpu
pdfminer.six
sentence-transformers
transformers
bitsandbytes
accelerate
torch
mlx-lm
# dev / tooling
pre-commit
ruff
